feat: добавляет реализацию избранных товаров для пользователей

+++ apps/product/cart_views.py
@@ -12,10 +12,9 @@ from apps.product.cart_serializers import (
     CartItemCreateSerializer,
     CartModelDictSerializer,
     CartModelSerializer,
-    FavModelDictSerializer,
     FavoriteCreateSerializer,
-    FavoriteListSerializer,
     FavoriteSerializer,
+    ShortProductSerializer,
 )
 from apps.product.models import CartItem, CartModel, Favorite, Product
 
@@ -96,19 +95,18 @@ def del_item(request, id):
     return Response(serializer.data, status=status.HTTP_200_OK)
 
 
-# Favorites
 @extend_schema(responses={status.HTTP_200_OK: FavoriteSerializer}, methods=['GET'])
 @api_view(['GET'])
 def fav_list(request, pk=None):
     """Возвращает данные о товарах в избранном пользователя."""
     user = request.user
     if user.is_authenticated:
-        favorites = Favorite.objects.filter(user=user)
-        serializer = FavoriteListSerializer({'user': user, 'products': favorites}, context={'request': request})
+        favorite_products = Product.objects.filter(favorites__user=user)
+        serializer = ShortProductSerializer(instance=favorite_products, many=True, context={'request': request})
         return Response(serializer.data)
-    favorites_req = CartAndFavorites(request=request)
-    favorites = favorites_req.extract_items_favorites()
-    serializer = FavModelDictSerializer(instance=favorites, context={'request': request})
+    cart = CartAndFavorites(request=request)
+    cart_items = cart.extract_items_favorites()
+    serializer = FavoriteSerializer(instance=cart_items, context={'request': request})
     return Response(serializer.data)
 
 
@@ -116,7 +114,7 @@ def fav_list(request, pk=None):
     request=FavoriteCreateSerializer,
     responses={
         status.HTTP_201_CREATED: OpenApiResponse(
-            response=FavoriteListSerializer, description='Успешное добавление товара в избранное'
+            response=FavoriteSerializer, description='Успешное добавление товара в избранное'
         )
     },
     methods=['POST'],
@@ -125,6 +123,7 @@ def fav_list(request, pk=None):
 def add_favorite(request):
     """Добавляет товар в избранное."""
     user = request.user
+    product = get_object_or_404(Product, pk=request.data['product'])
 
     if not user.is_authenticated:
         favorites = CartAndFavorites(request=request)
@@ -132,20 +131,21 @@ def add_favorite(request):
         serializer.is_valid(raise_exception=True)
         favorites.add_to_favorites(product_id=serializer.data.get('product'))
         fav_items = favorites.extract_items_favorites()
-        serializer = FavModelDictSerializer(instance=fav_items, context={'request': request})
+        serializer = FavoriteSerializer(instance=fav_items, context={'request': request})
         return Response(serializer.data, status=status.HTTP_201_CREATED)
-    serializer = FavoriteCreateSerializer(data=request.data, context={'request': request})
-    serializer.is_valid(raise_exception=True)
-    product = get_object_or_404(Product, pk=request.data['product'])
-    favorite, _ = Favorite.objects.get_or_create(user=user, product=product)
-    serializer = FavoriteListSerializer(instance=favorite, context={'request': request})
-
+    Favorite.objects.get_or_create(user=user, product=product)
+    favorite_products = Product.objects.filter(favorites__user=user)
+    serializer = ShortProductSerializer(instance=favorite_products, many=True, context={'request': request})
     return Response(serializer.data, status=status.HTTP_201_CREATED)
 
 
 @extend_schema(
     parameters=[OpenApiParameter('id', OpenApiTypes.INT, OpenApiParameter.PATH, description='Идентификатор продукта')],
-    responses={status.HTTP_200_OK: OpenApiResponse(description='Успешное удаление товара из избранного')},
+    responses={
+        status.HTTP_200_OK: OpenApiResponse(
+            response=FavoriteSerializer, description='Успешное удаление товара из избранного'
+        )
+    },
     methods=['DELETE'],
 )
 @api_view(['DELETE'])
@@ -155,13 +155,14 @@ def del_favorite(request, id):
     user = request.user
     if not user.is_authenticated:
         favorites = CartAndFavorites(request=request)
-        favorites.remove_from_favorites(product_id=product.id)
+        favorites.remove_from_favorites(product.id)
         fav_items = favorites.extract_items_favorites()
-        serializer = FavModelDictSerializer(instance=fav_items, context={'request': request})
+        serializer = FavoriteSerializer(instance=fav_items, context={'request': request})
         return Response(serializer.data, status=status.HTTP_200_OK)
 
-    favorite = get_object_or_404(Favorite, product=product, user=user)
-    favorite.delete()
-
-    serializer = FavoriteListSerializer(instance=user.favorites.all(), context={'request': request})
+    favorite = Favorite.objects.filter(user=user, product=product).first()
+    if favorite:
+        favorite.delete()
+    favorite_products = Product.objects.filter(favorites__user=user)
+    serializer = ShortProductSerializer(instance=favorite_products, many=True, context={'request': request})
     return Response(serializer.data, status=status.HTTP_200_OK)
